compiler: change rect representation to fix distances.
This is a good one.

So, previously, I was representing 2D rectangles as very thin 3D
rectangular solids.  "The thickness doesn't matter," I thought, "because
I'm just going to slice the distance field at the XY plane anyway."

This is, of course, wrong.  Given a rectangular solid of (say) 300x200x1
unit, the maximum of its interior distance field is 0.5!

And the *interior* distance field really starts to matter when you
negate a volume, such as when computing an intersection.

The net effect: algorithms using optimizations derived from assumptions
about Lipschitz continuity, such as sphere tracing, would stall in
processing the intersected volumes.  Think of the resulting field as a
sphere-tracer tarpit: no ray can move more than 0.5 units per hop!

I've now adjusted the Z dimension of this 2D primitive to be large
enough that it's immaterial for computations in the XY plane.  This
avoids the tarpit.