outline: add sparse marching squares.
This is about 10x faster at q=1 for the ex/box design.  More important,
though, is how it scales with q.

For the common laser-cutting case of long outlines separated by seas of
not-outline, time-cost scales linearly in q instead of exponentially
with the dense algorithm.  This makes higher q settings feasible.