Add 'main' submodules so 'raco test' passes.
The other option was to add a dummy 'test' submodule to each
command-line entry point.  But this would suppress elaboration of the
module, meaning we could miss typo'd symbols and such.

It seemed like a better idea to move the startup routine into a
function, which will be checked during 'raco test', and use a 'main'
submodule to separate startup from module evaluation.